separate out pytest-only tests, add to nose ignore lists
