+++ tests/tests.yaml
@@ -221,5 +221,7 @@ other_tests:
      - "--exclude-test=yt.frontends.gdf.tests.test_outputs.TestGDF"
      - "--exclude-test=yt.frontends.adaptahop.tests.test_outputs"
      - "--exclude-test=yt.frontends.stream.tests.test_stream_particles.test_stream_non_cartesian_particles"
+     - "--ignore-file=test_offaxisprojection_pytestonly\\.py"
+     - "--ignore-file=test_sph_pixelization_pytestonly\\.py"
   cookbook:
      - 'doc/source/cookbook/tests/test_cookbook.py'
